MNIST project has been added
In this project, there are two files, run.py and predict.py. We are using run.py in order to train the system, and then predict.py is for prediction as well. I put 4.jpg file inside, the code will be able to find out what digit we have in this way.